requirements.txt
Django==1.8.4
django-ckeditor==5.0.2
django-embed-video==1.0.0
djangocms-admin-style==0.2.8
Pillow==2.9.0
requests==2.7.0
sorl-thumbnail==12.3
